requirements.txt
Django
django-solo
gunicorn
Pillow
psycopg2
requests
raven
django-froala-editor